Split algorithm code and fix bug.
