+++ src/main/java/oripa/domain/fold/LayerOrderEnumerator.java
@@ -30,11 +30,11 @@ import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import java.util.stream.Stream;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import oripa.domain.fold.DeterministicLayerOrderEstimator.EstimationResult;
 import oripa.domain.fold.halfedge.OriEdge;
 import oripa.domain.fold.halfedge.OriFace;
 import oripa.domain.fold.halfedge.OriHalfedge;
@@ -55,48 +55,14 @@ import oripa.util.StopWatch;
 public class LayerOrderEnumerator {
 	private static final Logger logger = LoggerFactory.getLogger(LayerOrderEnumerator.class);
 
-	private HashSet<StackConditionOf4Faces> condition4s;
+//	private HashSet<StackConditionOf4Faces> condition4s;
 	private List<SubFace> subFaces;
-
-	/**
-	 * [faceID1][faceID2]
-	 */
-	private List<Integer>[][] overlappingFaceIndexIntersections;
-
-	/**
-	 * Key: halfedge, value: set of indices of faces which are under the
-	 * halfedge.
-	 */
-	private Map<OriHalfedge, Set<Integer>> faceIndicesOnHalfEdge;
 
 	private AtomicInteger callCount;
 	private AtomicInteger penetrationTestCallCount;
 	private AtomicInteger penetrationCount;
 
 	private final SubFacesFactory subFacesFactory;
-
-	private enum EstimationResult {
-		NOT_CHANGED(0),
-		CHANGED(1),
-		UNFOLDABLE(2);
-
-		private EstimationResult(final int order) {
-			this.order = order;
-		}
-
-		private int order;
-
-		private int getOrder() {
-			return order;
-		}
-
-		public static EstimationResult selectStronger(final EstimationResult a, final EstimationResult b) {
-			return Stream.of(a, b)
-					.sorted(Comparator.comparing(EstimationResult::getOrder).reversed())
-					.findFirst()
-					.get();
-		}
-	}
 
 	public LayerOrderEnumerator(final SubFacesFactory subFacesFactory) {
 		this.subFacesFactory = subFacesFactory;
@@ -112,24 +78,27 @@ public class LayerOrderEnumerator {
 		subFaces = subFacesFactory.createSubFaces(faces, paperSize, eps);
 		logger.debug("subFaces.size() = " + subFaces.size());
 
-		OverlapRelation overlapRelation = createOverlapRelation(faces, eps);
-
 		// Set overlap relations based on valley/mountain folds information
-		determineOverlapRelationByLineType(faces, overlapRelation);
+		OverlapRelation overlapRelation = new OverlapRelationFactory().createOverlapRelationByLineType(faces, eps);
 
 		var watch = new StopWatch(true);
 
-		overlappingFaceIndexIntersections = createOverlappingFaceIndexIntersections(faces, overlapRelation);
-		faceIndicesOnHalfEdge = createFaceIndicesOnHalfEdge(faces, eps);
-
 		logger.debug("preprocessing time = {}[ms]", watch.getMilliSec());
 
-		holdCondition3s(faces, overlapRelation);
+		var overlappingFaceIndexIntersections = createOverlappingFaceIndexIntersections(faces, overlapRelation);
+		var faceIndicesOnHalfEdge = createFaceIndicesOnHalfEdge(faces, eps);
+		holdCondition3s(faces, overlapRelation, overlappingFaceIndexIntersections, faceIndicesOnHalfEdge);
 
-		condition4s = new HashSet<>();
-		holdCondition4s(faces, edges, overlapRelation, eps);
+		var condition4s = holdCondition4s(faces, edges, overlapRelation, eps);
 
-		if (estimate(faces, overlapRelation) == EstimationResult.UNFOLDABLE) {
+		var estimator = new DeterministicLayerOrderEstimator(
+				faces, subFaces,
+				overlappingFaceIndexIntersections,
+				faceIndicesOnHalfEdge,
+				condition4s);
+		var estimationResult = estimator.estimate(overlapRelation, eps);
+
+		if (estimationResult == EstimationResult.UNFOLDABLE) {
 			logger.debug("found unfoldable before searching.");
 			return List.of();
 		}
@@ -177,23 +146,9 @@ public class LayerOrderEnumerator {
 		return overlapRelations;
 	}
 
-	private int countUndefinedRelations(final OverlapRelation overlapRelation) {
-		int size = overlapRelation.getSize();
-
-		int count = 0;
-		for (int i = 0; i < size; i++) {
-			for (int j = 0; j < size; j++) {
-				if (overlapRelation.isUndefined(i, j)) {
-					count++;
-				}
-			}
-		}
-
-		return count;
-	}
-
 	@SuppressWarnings("unchecked")
-	private List<Integer>[][] createOverlappingFaceIndexIntersections(final List<OriFace> faces,
+	private List<Integer>[][] createOverlappingFaceIndexIntersections(
+			final List<OriFace> faces,
 			final OverlapRelation overlapRelation) {
 		List<Set<Integer>> indices = IntStream.range(0, faces.size())
 				.mapToObj(i -> new HashSet<Integer>())
@@ -234,7 +189,8 @@ public class LayerOrderEnumerator {
 	}
 
 	private Map<OriHalfedge, Set<Integer>> createFaceIndicesOnHalfEdge(
-			final List<OriFace> faces, final double eps) {
+			final List<OriFace> faces,
+			final double eps) {
 
 		Map<OriHalfedge, Set<Integer>> indices = new HashMap<>();
 
@@ -259,6 +215,21 @@ public class LayerOrderEnumerator {
 		}
 
 		return indices;
+	}
+
+	private int countUndefinedRelations(final OverlapRelation overlapRelation) {
+		int size = overlapRelation.getSize();
+
+		int count = 0;
+		for (int i = 0; i < size; i++) {
+			for (int j = 0; j < size; j++) {
+				if (overlapRelation.isUndefined(i, j)) {
+					count++;
+				}
+			}
+		}
+
+		return count;
 	}
 
 	/**
@@ -328,45 +299,6 @@ public class LayerOrderEnumerator {
 	}
 
 	/**
-	 * Determines overlap relations by necessary conditions.
-	 *
-	 * @param faces
-	 *            all faces.
-	 * @param overlapRelation
-	 *            overlap relation matrix
-	 */
-	private EstimationResult estimate(final List<OriFace> faces, final OverlapRelation overlapRelation) {
-		int estimationLoopCount = 0;
-
-		var watch = new StopWatch(true);
-		boolean changed;
-		do {
-			changed = false;
-			var faceCoverResult = estimateBy3FaceCover(faces, overlapRelation);
-			if (faceCoverResult == EstimationResult.UNFOLDABLE) {
-				return EstimationResult.UNFOLDABLE;
-			}
-			changed |= faceCoverResult == EstimationResult.CHANGED;
-
-			changed |= estimateBy3FaceTransitiveRelation(overlapRelation);
-
-			faceCoverResult = estimateBy4FaceStackCondition(overlapRelation);
-			if (faceCoverResult == EstimationResult.UNFOLDABLE) {
-				return EstimationResult.UNFOLDABLE;
-			}
-
-			estimationLoopCount++;
-		} while (changed);
-		logger.debug("#estimation = {}", estimationLoopCount);
-		logger.debug("estimation time {}[ms]", watch.getMilliSec());
-
-		if (changed) {
-			return EstimationResult.CHANGED;
-		}
-		return EstimationResult.NOT_CHANGED;
-	}
-
-	/**
 	 * Creates 3-face condition and sets to subfaces: If face[i] and face[j]
 	 * touching edge are covered by face[k] then OR[i][k] = OR[j][k]
 	 *
@@ -376,7 +308,9 @@ public class LayerOrderEnumerator {
 	 *            overlap relation matrix
 	 */
 	private void holdCondition3s(
-			final List<OriFace> faces, final OverlapRelation overlapRelation) {
+			final List<OriFace> faces, final OverlapRelation overlapRelation,
+			final List<Integer>[][] overlappingFaceIndexIntersections,
+			final Map<OriHalfedge, Set<Integer>> faceIndicesOnHalfEdge) {
 
 		int conditionCount = 0;
 
@@ -443,8 +377,9 @@ public class LayerOrderEnumerator {
 	 * @param overlapRelation
 	 *            overlap relation matrix
 	 */
-	private void holdCondition4s(final List<OriFace> faces,
+	private Set<StackConditionOf4Faces> holdCondition4s(final List<OriFace> faces,
 			final List<OriEdge> edges, final OverlapRelation overlapRelation, final double eps) {
+		var condition4s = new HashSet<StackConditionOf4Faces>();
 
 		int edgeNum = edges.size();
 		logger.debug("edgeNum = " + edgeNum);
@@ -521,6 +456,7 @@ public class LayerOrderEnumerator {
 		logger.debug("#condition4 = {}", condition4s.size());
 		logger.debug("condition4s computation time {}[ms]", watch.getMilliSec());
 
+		return condition4s;
 	}
 
 	private Map<OriFace, Set<SubFace>> createSubFacesOfEachFace(final List<OriFace> faces) {
@@ -535,343 +471,5 @@ public class LayerOrderEnumerator {
 		}
 
 		return subFacesOfEachFace;
-	}
-
-	/**
-	 * Determines overlap relation using 4-face condition.
-	 *
-	 * @param overlapRelation
-	 * @return
-	 */
-	private EstimationResult estimateBy4FaceStackCondition(final OverlapRelation overlapRelation) {
-
-		var changed = EstimationResult.NOT_CHANGED;
-
-		for (StackConditionOf4Faces cond : condition4s) {
-
-			// if: lower1 > upper2, then: upper1 > upper2, upper1 > lower2,
-			// lower1 > lower2
-			if (overlapRelation.isLower(cond.lower1, cond.upper2)) {
-				var result = setLowerIfPossible(overlapRelation, cond.upper1, cond.upper2);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.upper1, cond.lower2);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.lower1, cond.lower2);
-				changed = EstimationResult.selectStronger(result, changed);
-			}
-
-			// if: lower2 > upper1, then: upper2 > upper1, upper2 > lower1,
-			// lower2 > lower1
-			if (overlapRelation.isLower(cond.lower2, cond.upper1)) {
-				var result = setLowerIfPossible(overlapRelation, cond.upper2, cond.upper1);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.upper2, cond.lower1);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.lower2, cond.lower1);
-				changed = EstimationResult.selectStronger(result, changed);
-			}
-
-			// if: upper1 > upper2 > lower1, then: upper1 > lower2, lower2 >
-			// lower1
-			if (overlapRelation.isLower(cond.upper1, cond.upper2)
-					&& overlapRelation.isLower(cond.upper2, cond.lower1)) {
-				var result = setLowerIfPossible(overlapRelation, cond.upper1, cond.lower2);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.lower2, cond.lower1);
-				changed = EstimationResult.selectStronger(result, changed);
-			}
-
-			// if: upper1 > lower2 > lower1, then: upper1 > upper2, upper2 >
-			// lower1
-			if (overlapRelation.isLower(cond.upper1, cond.lower2)
-					&& overlapRelation.isLower(cond.lower2, cond.lower1)) {
-				var result = setLowerIfPossible(overlapRelation, cond.upper1, cond.upper2);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.upper2, cond.lower1);
-				changed = EstimationResult.selectStronger(result, changed);
-			}
-
-			// if: upper2 > upper1 > lower2, then: upper2 > lower1, lower1 >
-			// lower2
-			if (overlapRelation.isLower(cond.upper2, cond.upper1)
-					&& overlapRelation.isLower(cond.upper1, cond.lower2)) {
-				var result = setLowerIfPossible(overlapRelation, cond.upper2, cond.lower1);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.lower1, cond.lower2);
-				changed = EstimationResult.selectStronger(result, changed);
-			}
-
-			// if: upper2 > lower1 > lower2, then: upper2 > upper1, upper1 >
-			// lower2
-			if (overlapRelation.isLower(cond.upper2, cond.lower1)
-					&& overlapRelation.isLower(cond.lower1, cond.lower2)) {
-				var result = setLowerIfPossible(overlapRelation, cond.upper2, cond.upper1);
-				changed = EstimationResult.selectStronger(result, changed);
-
-				result = setLowerIfPossible(overlapRelation, cond.upper1, cond.lower2);
-				changed = EstimationResult.selectStronger(result, changed);
-			}
-
-			if (changed == EstimationResult.UNFOLDABLE) {
-				return EstimationResult.UNFOLDABLE;
-			}
-
-		}
-
-		return changed;
-	}
-
-	private EstimationResult setLowerIfPossible(final OverlapRelation overlapRelation, final int i, final int j) {
-
-		if (overlapRelation.setLowerIfUndefined(i, j)) {
-			return EstimationResult.CHANGED;
-		}
-		if (overlapRelation.isUpper(i, j)) {
-			// conflict.
-			return EstimationResult.UNFOLDABLE;
-		}
-
-		return EstimationResult.NOT_CHANGED;
-	}
-
-	/**
-	 * If the subface a>b and b>c then a>c
-	 *
-	 * @param overlapRelation
-	 *            overlap-relation matrix
-	 * @return whether overlapRelation is changed or not.
-	 */
-	private boolean estimateBy3FaceTransitiveRelation(final OverlapRelation overlapRelation) {
-		boolean changed = false;
-
-		for (SubFace sub : subFaces) {
-			while (updateOverlapRelationBy3FaceTransitiveRelation(sub, overlapRelation)) {
-				changed = true;
-			}
-		}
-		return changed;
-	}
-
-	/**
-	 * Updates {@code overlapRelation} by 3-face stack condition.
-	 *
-	 * @param sub
-	 *            subface.
-	 * @param overlapRelation
-	 *            overlap relation matrix.
-	 * @return true if an update happens.
-	 */
-	private boolean updateOverlapRelationBy3FaceTransitiveRelation(final SubFace sub,
-			final OverlapRelation overlapRelation) {
-
-		for (int i = 0; i < sub.getParentFaceCount(); i++) {
-			for (int j = i + 1; j < sub.getParentFaceCount(); j++) {
-
-				// search for undetermined relations
-				int index_i = sub.getParentFace(i).getFaceID();
-				int index_j = sub.getParentFace(j).getFaceID();
-
-				if (overlapRelation.isNoOverlap(index_i, index_j)) {
-					continue;
-				}
-				if (!overlapRelation.isUndefined(index_i, index_j)) {
-					continue;
-				}
-				// Find the intermediary face
-				for (int k = 0; k < sub.getParentFaceCount(); k++) {
-					if (k == i || k == j) {
-						continue;
-					}
-
-					int index_k = sub.getParentFace(k).getFaceID();
-
-					if (overlapRelation.isUpper(index_i, index_k)
-							&& overlapRelation.isUpper(index_k, index_j)) {
-						overlapRelation.setUpper(index_i, index_j);
-						return true;
-					}
-					if (overlapRelation.isLower(index_i, index_k)
-							&& overlapRelation.isLower(index_k, index_j)) {
-						overlapRelation.setLower(index_i, index_j);
-						return true;
-					}
-				}
-			}
-		}
-		return false;
-	}
-
-	/**
-	 * If face[i] and face[j] touching edge is covered by face[k] then
-	 * overlapRelation[i][k] = overlapRelation[j][k].
-	 *
-	 * @param faces
-	 *            all faces of the model
-	 * @param overlapRelation
-	 *            overlap relation matrix
-	 * @return whether overlapRelation is changed or not.
-	 */
-	private EstimationResult estimateBy3FaceCover(
-			final List<OriFace> faces,
-			final OverlapRelation overlapRelation) {
-
-		EstimationResult changed = EstimationResult.NOT_CHANGED;
-		for (OriFace f_i : faces) {
-			var updateResult = updateBy3FaceCover(f_i, overlapRelation);
-			switch (updateResult) {
-			case UNFOLDABLE:
-				return EstimationResult.UNFOLDABLE;
-			case CHANGED:
-				changed = EstimationResult.CHANGED;
-			default:
-				break;
-			}
-		}
-
-		return changed;
-
-	}
-
-	private EstimationResult updateBy3FaceCover(
-			final OriFace f_i,
-			final OverlapRelation overlapRelation) {
-		int index_i = f_i.getFaceID();
-
-		var changed = EstimationResult.NOT_CHANGED;
-		for (OriHalfedge he : f_i.halfedgeIterable()) {
-			var pair = he.getPair();
-			if (pair == null) {
-				continue;
-			}
-			int index_j = pair.getFace().getFaceID();
-
-			var indices = overlappingFaceIndexIntersections[index_i][index_j];
-			for (var index_k : indices) {
-				if (index_i == index_k || index_j == index_k) {
-					continue;
-				}
-				if (!faceIndicesOnHalfEdge.get(he).contains(index_k)) {
-					continue;
-				}
-				if (!overlapRelation.isUndefined(index_i, index_k)) {
-					var result = setIfPossible(
-							overlapRelation, index_j, index_k, overlapRelation.get(index_i, index_k));
-					changed = EstimationResult.selectStronger(result, changed);
-				}
-				if (!overlapRelation.isUndefined(index_j, index_k)) {
-					var result = setIfPossible(
-							overlapRelation, index_i, index_k, overlapRelation.get(index_j, index_k));
-					changed = EstimationResult.selectStronger(result, changed);
-				}
-
-				if (changed == EstimationResult.UNFOLDABLE) {
-					return EstimationResult.UNFOLDABLE;
-				}
-
-			}
-		}
-
-		return changed;
-
-	}
-
-	private EstimationResult setIfPossible(final OverlapRelation overlapRelation,
-			final int i, final int j, final byte value) {
-
-		if (overlapRelation.setIfUndefined(i, j, value)) {
-			return EstimationResult.CHANGED;
-		}
-		if (overlapRelation.get(i, j) != value) {
-			// conflict.
-			return EstimationResult.UNFOLDABLE;
-		}
-
-		return EstimationResult.NOT_CHANGED;
-	}
-
-	/**
-	 * Determines the overlap relations by mountain/valley.
-	 *
-	 * @param faces
-	 *            all faces of the model
-	 * @param overlapRelation
-	 *            overlap relation matrix
-	 */
-	private void determineOverlapRelationByLineType(
-			final List<OriFace> faces, final OverlapRelation overlapRelation) {
-
-		for (OriFace face : faces) {
-			for (OriHalfedge he : face.halfedgeIterable()) {
-				var pair = he.getPair();
-				if (pair == null) {
-					continue;
-				}
-				OriFace pairFace = pair.getFace();
-				var faceID = face.getFaceID();
-				var pairFaceID = pairFace.getFaceID();
-
-				// If the relation is already decided, skip
-				if (overlapRelation.isUpper(faceID, pairFaceID)
-						|| overlapRelation.isLower(faceID, pairFaceID)) {
-					continue;
-				}
-
-				var edge = he.getEdge();
-				if ((face.isFaceFront() && edge.isMountain())
-						|| (!face.isFaceFront() && edge.isValley())) {
-					overlapRelation.setUpper(faceID, pairFaceID);
-				} else {
-					overlapRelation.setLower(faceID, pairFaceID);
-				}
-			}
-		}
-	}
-
-	/**
-	 * creates the matrix overlapRelation and fills it with "no overlap" or
-	 * "undefined"
-	 *
-	 * @param faces
-	 *            all faces of the model
-	 * @param paperSize
-	 *            paper size before fold
-	 * @return initialized overlap relation matrix
-	 */
-	private OverlapRelation createOverlapRelation(final List<OriFace> faces, final double eps) {
-
-		int size = faces.size();
-		OverlapRelation overlapRelation = new OverlapRelation(size);
-
-		int countOfZeros = 0;
-		for (int i = 0; i < size; i++) {
-			overlapRelation.setNoOverlap(i, i);
-			countOfZeros++;
-			for (int j = i + 1; j < size; j++) {
-				if (OriGeomUtil.isFaceOverlap(faces.get(i), faces.get(j), eps)) {
-					overlapRelation.setUndefined(i, j);
-				} else {
-					overlapRelation.setNoOverlap(i, j);
-					countOfZeros += 2;
-				}
-			}
-		}
-
-		double rate = ((double) countOfZeros) / (size * size);
-		logger.debug("sparsity of overlap relation matrix = {}", rate);
-		// One element in dictionary of keys for byte value needs at least 16
-		// bytes.
-		if (rate > 15.0 / 16) {
-			logger.debug("use sparse matrix for overlap relation.");
-			overlapRelation.switchToSparseMatrix();
-		}
-
-		return overlapRelation;
 	}
 }
